+++ admin-ui/src/pages/admin/tenants/[id].tsx
@@ -14,6 +14,11 @@ import PostHogIntegrationTab from "./PostHogIntegrationTab";
 import Ga4IntegrationTab from "./Ga4IntegrationTab";
 import ApiKeysTab, { fetchApiKeys } from "./ApiKeysTab";
 import EmbedCodeTab from "./EmbedCodeTab";
+import DeepResearchTab from "./DeepResearchTab";
+import ConversionTypesTab from "./ConversionTypesTab";
+import AnalyticsSummaryTab from "./AnalyticsSummaryTab";
+import BillingInfoTab from "./BillingInfoTab";
+import NotificationPreferencesTab from "./NotificationPreferencesTab";
 import type { TenantFeatures, TenantDetail, ApiKey } from "./types";
 import { CARD_STYLE } from "./types";
 
@@ -118,33 +123,6 @@ async function updateAvatarSettings(
     lemonslice_agent_id: json.lemonslice_agent_id ?? null,
   } as TenantDetail;
 }
-
-async function updateDeepResearchSettings(
-  tenantId: string,
-  deepResearch: boolean,
-  currentFeatures: TenantFeatures
-): Promise<TenantDetail> {
-  const features: TenantFeatures = { ...currentFeatures, deep_research: deepResearch };
-  const res = await authFetch(`${API_BASE}/v1/admin/tenants/${tenantId}`, {
-    method: "PATCH",
-    body: JSON.stringify({ features }),
-  });
-  if (!res.ok) throw new Error(`HTTP ${res.status}`);
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const raw = (await res.json()) as any;
-  const json = "tenant" in raw ? raw.tenant : raw;
-  return {
-    ...json,
-    status: json.is_active ? "active" : "inactive",
-    allowed_origins: json.allowed_origins ?? [],
-    billing_enabled: json.billing_enabled ?? false,
-    billing_free_from: json.billing_free_from ?? null,
-    billing_free_until: json.billing_free_until ?? null,
-    features: json.features ?? { avatar: false, voice: false, rag: true, deep_research: false },
-    lemonslice_agent_id: json.lemonslice_agent_id ?? null,
-  } as TenantDetail;
-}
-
 
 async function updateBilling(
   tenantId: string,
@@ -801,607 +779,6 @@ function SettingsTab({
         <BillingSection tenant={tenant} onUpdate={onBillingUpdate} />
       )}
     </form>
-  );
-}
-
-// ─── ディープリサーチ設定タブ ─────────────────────────────────────────────────
-
-function DeepResearchTab({
-  tenant,
-  onUpdate,
-  showToast,
-}: {
-  tenant: TenantDetail;
-  onUpdate: (updated: TenantDetail) => void;
-  showToast: (msg: string) => void;
-}) {
-  const [deepResearch, setDeepResearch] = useState<boolean>(tenant.features.deep_research ?? false);
-  const [saving, setSaving] = useState(false);
-  const [confirmPending, setConfirmPending] = useState(false);
-
-  const handleToggle = async () => {
-    const next = !deepResearch;
-    if (next) {
-      // ON切り替え → 確認ダイアログ
-      setConfirmPending(true);
-      return;
-    }
-    // OFF切り替え → 即座に反映
-    await save(false);
-  };
-
-  const save = async (value: boolean) => {
-    setSaving(true);
-    try {
-      const updated = await updateDeepResearchSettings(tenant.id, value, tenant.features);
-      setDeepResearch(value);
-      onUpdate(updated);
-      showToast("✅ 設定を保存しました");
-    } catch {
-      showToast("❌ 保存に失敗しました。もう一度お試しください");
-    } finally {
-      setSaving(false);
-      setConfirmPending(false);
-    }
-  };
-
-  return (
-    <div style={{ display: "flex", flexDirection: "column", gap: 20, maxWidth: "100%", overflowX: "hidden", boxSizing: "border-box" }}>
-      {/* 確認ダイアログ */}
-      {confirmPending && (
-        <div
-          style={{
-            position: "fixed", inset: 0, zIndex: 1000,
-            background: "rgba(0,0,0,0.7)",
-            display: "flex", alignItems: "center", justifyContent: "center",
-            padding: 20,
-          }}
-        >
-          <div
-            style={{
-              background: "#111827", borderRadius: 14,
-              border: "1px solid #374151",
-              padding: "28px 24px", maxWidth: 400, width: "100%",
-            }}
-          >
-            <p style={{ color: "#e5e7eb", fontSize: 15, fontWeight: 600, margin: "0 0 12px" }}>
-              ディープリサーチをONにしますか？
-            </p>
-            <p style={{ color: "#9ca3af", fontSize: 13, margin: "0 0 24px", lineHeight: 1.6 }}>
-              ディープリサーチをONにすると、追加コスト（月$3〜8程度）が発生します。よろしいですか？
-            </p>
-            <div style={{ display: "flex", gap: 12, justifyContent: "flex-end" }}>
-              <button
-                type="button"
-                onClick={() => setConfirmPending(false)}
-                style={{
-                  padding: "10px 18px", borderRadius: 8,
-                  border: "1px solid #374151", background: "transparent",
-                  color: "#9ca3af", fontSize: 14, cursor: "pointer",
-                }}
-              >
-                キャンセル
-              </button>
-              <button
-                type="button"
-                onClick={() => save(true)}
-                disabled={saving}
-                style={{
-                  padding: "10px 18px", borderRadius: 8,
-                  border: "none", background: saving ? "#1f2937" : "#1d4ed8",
-                  color: saving ? "#6b7280" : "#fff",
-                  fontSize: 14, fontWeight: 700, cursor: saving ? "not-allowed" : "pointer",
-                }}
-              >
-                {saving ? "保存中..." : "ONにする"}
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* メインカード */}
-      <div
-        style={{
-          ...CARD_STYLE,
-          display: "flex", alignItems: "center", justifyContent: "space-between",
-          flexWrap: "wrap", gap: 16,
-        }}
-      >
-        <div style={{ flex: 1, minWidth: 0 }}>
-          <p style={{ margin: "0 0 4px", fontWeight: 600, color: "#e5e7eb", fontSize: 15 }}>
-            🔬 ディープリサーチ（AI提案の精度向上）
-          </p>
-          <p style={{ margin: 0, fontSize: 13, color: "#9ca3af" }}>
-            AIの改善提案に最新の市場動向・心理学研究を反映します
-          </p>
-        </div>
-        <button
-          type="button"
-          onClick={handleToggle}
-          disabled={saving}
-          aria-label="ディープリサーチ切り替え"
-          style={{
-            position: "relative",
-            display: "inline-flex", alignItems: "center",
-            width: 56, height: 32, borderRadius: 16,
-            border: "none",
-            background: deepResearch ? "#2563eb" : "#374151",
-            cursor: saving ? "not-allowed" : "pointer",
-            transition: "background 0.2s", flexShrink: 0,
-            opacity: saving ? 0.6 : 1,
-          }}
-        >
-          <span
-            style={{
-              display: "inline-block", width: 24, height: 24, borderRadius: "50%",
-              background: "#fff",
-              transform: deepResearch ? "translateX(28px)" : "translateX(4px)",
-              transition: "transform 0.2s",
-            }}
-          />
-        </button>
-      </div>
-
-      {/* ONにすると何が実現できるか */}
-      <div
-        style={{
-          borderRadius: 12,
-          border: "1px solid rgba(96,165,250,0.3)",
-          background: "rgba(29,78,216,0.1)",
-          padding: "16px 18px",
-        }}
-      >
-        <p style={{ margin: "0 0 8px", fontWeight: 600, color: "#93c5fd", fontSize: 13 }}>
-          ONにすると：
-        </p>
-        <ul style={{ margin: 0, paddingLeft: 0, listStyle: "none", display: "flex", flexDirection: "column", gap: 4 }}>
-          {[
-            "チューニングルール提案に最新の業界動向が反映されます",
-            "ナレッジの穴の推薦精度が向上します",
-            "管理画面AIアシスタントが外部の知見も参照します",
-          ].map((item, i) => (
-            <li key={i} style={{ color: "#bfdbfe", fontSize: 13 }}>・{item}</li>
-          ))}
-        </ul>
-      </div>
-
-      {/* コスト説明 */}
-      <div
-        style={{
-          borderRadius: 10,
-          border: "1px solid #1f2937",
-          background: "rgba(17,24,39,0.5)",
-          padding: "12px 16px",
-          fontSize: 12, color: "#6b7280", lineHeight: 1.6,
-        }}
-      >
-        <p style={{ margin: "0 0 4px" }}>💰 コスト目安：月あたり約 $3〜8 の追加（提案1回あたり約 $0.05〜0.10）</p>
-        <p style={{ margin: 0 }}>※ 通常の提案機能は無料で引き続きご利用いただけます</p>
-      </div>
-
-      <p style={{ margin: 0, fontSize: 12, color: "#4b5563", textAlign: "right" }}>
-        現在: <strong style={{ color: deepResearch ? "#60a5fa" : "#6b7280" }}>{deepResearch ? "ON" : "OFF"}</strong>
-      </p>
-    </div>
-  );
-}
-
-// ─── コンバージョンタイプ設定タブ ─────────────────────────────────────────────
-
-function ConversionTypesTab({
-  tenant,
-  onUpdate,
-}: {
-  tenant: TenantDetail;
-  onUpdate: (updated: TenantDetail) => void;
-}) {
-  const [types, setTypes] = useState<string[]>(tenant.conversion_types ?? ["購入完了", "予約完了", "問い合わせ送信", "離脱", "不明"]);
-  const [newType, setNewType] = useState("");
-  const [saving, setSaving] = useState(false);
-  const [saveMsg, setSaveMsg] = useState<{ msg: string; ok: boolean } | null>(null);
-
-  const showMsg = (msg: string, ok: boolean) => {
-    setSaveMsg({ msg, ok });
-    setTimeout(() => setSaveMsg(null), 3000);
-  };
-
-  const addType = () => {
-    const trimmed = newType.trim();
-    if (!trimmed) return;
-    if (trimmed.length > 50) { showMsg("50文字以内で入力してください", false); return; }
-    if (types.includes(trimmed)) { showMsg("同じタイプがすでに存在します", false); return; }
-    if (types.length >= 10) { showMsg("最大10件まで登録できます", false); return; }
-    setTypes([...types, trimmed]);
-    setNewType("");
-  };
-
-  const removeType = (t: string) => setTypes(types.filter((x) => x !== t));
-
-  const handleSave = async () => {
-    if (types.length === 0) { showMsg("少なくとも1件必要です", false); return; }
-    setSaving(true);
-    try {
-      const res = await authFetch(`${API_BASE}/v1/admin/tenants/${tenant.id}`, {
-        method: "PATCH",
-        body: JSON.stringify({ conversion_types: types }),
-      });
-      if (!res.ok) throw new Error(`HTTP ${res.status}`);
-      const raw = (await res.json()) as any;
-      const json = "tenant" in raw ? raw.tenant : raw;
-      onUpdate({
-        ...json,
-        status: json.is_active ? "active" : "inactive",
-        allowed_origins: json.allowed_origins ?? [],
-        billing_enabled: json.billing_enabled ?? false,
-        billing_free_from: json.billing_free_from ?? null,
-        billing_free_until: json.billing_free_until ?? null,
-        features: json.features ?? { avatar: false, voice: false, rag: true },
-        lemonslice_agent_id: json.lemonslice_agent_id ?? null,
-        conversion_types: json.conversion_types ?? types,
-      } as TenantDetail);
-      showMsg("✅ コンバージョンタイプを保存しました", true);
-    } catch {
-      showMsg("保存に失敗しました", false);
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  return (
-    <div>
-      <p style={{ fontSize: 14, color: "#9ca3af", marginBottom: 20 }}>
-        お客様の行動結果のカテゴリを設定します。会話詳細ページでこのカテゴリを選んで成果を記録できます。
-      </p>
-      {saveMsg && (
-        <div style={{ marginBottom: 16, padding: "10px 16px", borderRadius: 8, fontSize: 14, fontWeight: 600,
-          background: saveMsg.ok ? "rgba(5,46,22,0.5)" : "rgba(127,29,29,0.4)",
-          border: `1px solid ${saveMsg.ok ? "rgba(74,222,128,0.3)" : "rgba(248,113,113,0.3)"}`,
-          color: saveMsg.ok ? "#86efac" : "#fca5a5",
-        }}>
-          {saveMsg.msg}
-        </div>
-      )}
-      {/* タイプ一覧 */}
-      <div style={{ display: "flex", flexWrap: "wrap", gap: 8, marginBottom: 20 }}>
-        {types.map((t) => (
-          <span key={t} style={{
-            display: "inline-flex", alignItems: "center", gap: 6,
-            padding: "6px 12px", borderRadius: 999, fontSize: 14,
-            background: "rgba(59,130,246,0.15)", border: "1px solid rgba(59,130,246,0.3)", color: "#93c5fd",
-          }}>
-            {t}
-            <button
-              onClick={() => removeType(t)}
-              style={{ background: "none", border: "none", color: "#6b7280", cursor: "pointer", fontSize: 14, padding: 0, lineHeight: 1 }}
-              title="削除"
-            >
-              ×
-            </button>
-          </span>
-        ))}
-        {types.length === 0 && <span style={{ fontSize: 14, color: "#6b7280" }}>タイプが登録されていません</span>}
-      </div>
-      {/* 追加フォーム */}
-      <div style={{ display: "flex", gap: 8, marginBottom: 20 }}>
-        <input
-          type="text"
-          value={newType}
-          onChange={(e) => setNewType(e.target.value)}
-          onKeyDown={(e) => { if (e.key === "Enter") { e.preventDefault(); addType(); } }}
-          placeholder="新しいタイプを追加（例: 資料請求）"
-          maxLength={50}
-          style={{
-            flex: 1, padding: "10px 14px", borderRadius: 8,
-            border: "1px solid #374151", background: "rgba(255,255,255,0.05)",
-            color: "#f9fafb", fontSize: 14,
-          }}
-        />
-        <button
-          onClick={addType}
-          disabled={!newType.trim() || types.length >= 10}
-          style={{
-            padding: "0 18px", minHeight: 44, borderRadius: 8,
-            border: "1px solid rgba(59,130,246,0.4)", background: "rgba(59,130,246,0.15)",
-            color: "#93c5fd", fontSize: 14, fontWeight: 600, cursor: "pointer", whiteSpace: "nowrap",
-          }}
-        >
-          ＋ 追加
-        </button>
-      </div>
-      <p style={{ fontSize: 12, color: "#6b7280", marginBottom: 20 }}>最大10件、各50文字以内。現在 {types.length}/10 件</p>
-      <button
-        onClick={handleSave}
-        disabled={saving}
-        style={{
-          padding: "12px 24px", minHeight: 48, borderRadius: 10,
-          border: "none", background: saving ? "#1f2937" : "#1d4ed8",
-          color: saving ? "#6b7280" : "#fff", fontSize: 15, fontWeight: 700, cursor: saving ? "not-allowed" : "pointer",
-        }}
-      >
-        {saving ? "保存中..." : "保存"}
-      </button>
-    </div>
-  );
-}
-
-// ─── タブ: アナリティクスサマリー ──────────────────────────────────────────────
-
-interface AnalyticsSummary {
-  period: string;
-  conversations: { total: number; avg_per_day: number };
-  cv: {
-    macro: { r2c_db: number; ga4: number; posthog: number; ranked_a: number; ranked_d: number };
-    micro: { r2c_db: number; ga4: number; posthog: number };
-  };
-  llm_usage: { tokens: number; cost_jpy: number; generations: number } | null;
-  alerts: { source_mismatch_count: number; ranked_d_count: number };
-}
-
-function AnalyticsSummaryTab({ tenantId }: { tenantId: string }) {
-  const [period, setPeriod] = useState<"last_7d" | "last_30d" | "last_90d">("last_30d");
-  const [data, setData] = useState<AnalyticsSummary | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-
-  useEffect(() => {
-    const load = async () => {
-      setLoading(true);
-      setError(null);
-      try {
-        const res = await authFetch(`${API_BASE}/v1/admin/tenants/${tenantId}/analytics-summary?period=${period}`);
-        if (!res.ok) throw new Error(`HTTP ${res.status}`);
-        setData(await res.json() as AnalyticsSummary);
-      } catch {
-        setError("データ取得に失敗しました");
-      } finally {
-        setLoading(false);
-      }
-    };
-    void load();
-  }, [tenantId, period]);
-
-  const periodLabel: Record<string, string> = { last_7d: "7日間", last_30d: "30日間", last_90d: "90日間" };
-
-  return (
-    <div style={{ display: "flex", flexDirection: "column", gap: 20 }}>
-      {/* Period selector */}
-      <div style={{ ...CARD_STYLE, display: "flex", gap: 10, alignItems: "center" }}>
-        <span style={{ fontSize: 14, color: "#9ca3af", fontWeight: 600 }}>期間:</span>
-        {(["last_7d", "last_30d", "last_90d"] as const).map((p) => (
-          <button
-            key={p}
-            type="button"
-            onClick={() => setPeriod(p)}
-            style={{
-              padding: "8px 16px",
-              minHeight: 36,
-              borderRadius: 8,
-              border: period === p ? "1px solid #4ade80" : "1px solid #374151",
-              background: period === p ? "rgba(34,197,94,0.15)" : "rgba(0,0,0,0.3)",
-              color: period === p ? "#4ade80" : "#9ca3af",
-              fontSize: 13,
-              fontWeight: 600,
-              cursor: "pointer",
-            }}
-          >
-            {periodLabel[p]}
-          </button>
-        ))}
-      </div>
-
-      {loading && <div style={{ color: "#6b7280", textAlign: "center", padding: 32 }}>読み込み中...</div>}
-      {error && <div style={{ color: "#f87171", padding: 16 }}>{error}</div>}
-
-      {data && !loading && (
-        <>
-          {/* Conversations */}
-          <div style={{ ...CARD_STYLE }}>
-            <h3 style={{ fontSize: 15, fontWeight: 700, color: "#d1d5db", margin: "0 0 16px" }}>💬 会話数</h3>
-            <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: 12 }}>
-              {[
-                { label: "総会話数", value: data.conversations.total.toLocaleString() },
-                { label: "1日平均", value: `${data.conversations.avg_per_day}件` },
-              ].map(({ label, value }) => (
-                <div key={label} style={{ padding: "16px", borderRadius: 10, background: "rgba(255,255,255,0.03)", border: "1px solid #1f2937" }}>
-                  <div style={{ fontSize: 12, color: "#9ca3af", marginBottom: 6 }}>{label}</div>
-                  <div style={{ fontSize: 22, fontWeight: 700, color: "#e5e7eb" }}>{value}</div>
-                </div>
-              ))}
-            </div>
-          </div>
-
-          {/* CV */}
-          <div style={{ ...CARD_STYLE }}>
-            <h3 style={{ fontSize: 15, fontWeight: 700, color: "#d1d5db", margin: "0 0 16px" }}>🎯 コンバージョン</h3>
-            <div style={{ display: "grid", gap: 10 }}>
-              {[
-                { label: "マクロCV (r2c_db)", value: data.cv.macro.r2c_db },
-                { label: "マクロCV (GA4)", value: data.cv.macro.ga4 },
-                { label: "マクロCV (PostHog)", value: data.cv.macro.posthog },
-                { label: "マイクロCV (r2c_db)", value: data.cv.micro.r2c_db },
-                { label: "マイクロCV (GA4)", value: data.cv.micro.ga4 },
-                { label: "ランクA (3ソース確認済)", value: data.cv.macro.ranked_a },
-              ].map(({ label, value }) => (
-                <div key={label} style={{ display: "flex", justifyContent: "space-between", padding: "10px 14px", borderRadius: 8, background: "rgba(255,255,255,0.02)", border: "1px solid #1f2937", fontSize: 14 }}>
-                  <span style={{ color: "#9ca3af" }}>{label}</span>
-                  <span style={{ color: "#e5e7eb", fontWeight: 600 }}>{value}</span>
-                </div>
-              ))}
-            </div>
-          </div>
-
-          {/* LLM Usage */}
-          {data.llm_usage && (
-            <div style={{ ...CARD_STYLE }}>
-              <h3 style={{ fontSize: 15, fontWeight: 700, color: "#d1d5db", margin: "0 0 16px" }}>🤖 LLM使用量（今月）</h3>
-              <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr 1fr", gap: 10 }}>
-                {[
-                  { label: "総トークン", value: data.llm_usage.tokens.toLocaleString() },
-                  { label: "推定コスト", value: `¥${data.llm_usage.cost_jpy.toLocaleString()}` },
-                  { label: "生成回数", value: data.llm_usage.generations.toLocaleString() },
-                ].map(({ label, value }) => (
-                  <div key={label} style={{ padding: "14px", borderRadius: 10, background: "rgba(255,255,255,0.03)", border: "1px solid #1f2937" }}>
-                    <div style={{ fontSize: 11, color: "#9ca3af", marginBottom: 4 }}>{label}</div>
-                    <div style={{ fontSize: 18, fontWeight: 700, color: "#4ade80" }}>{value}</div>
-                  </div>
-                ))}
-              </div>
-            </div>
-          )}
-
-          {/* Alerts */}
-          {(data.alerts.source_mismatch_count > 0 || data.alerts.ranked_d_count > 0) && (
-            <div style={{ ...CARD_STYLE, border: "1px solid rgba(239,68,68,0.4)", background: "rgba(127,29,29,0.15)" }}>
-              <h3 style={{ fontSize: 15, fontWeight: 700, color: "#f87171", margin: "0 0 12px" }}>⚠️ アラート</h3>
-              <div style={{ display: "grid", gap: 8 }}>
-                {data.alerts.source_mismatch_count > 0 && (
-                  <div style={{ fontSize: 14, color: "#fca5a5" }}>ソース不一致: {data.alerts.source_mismatch_count}件（同一イベントが複数ソースで記録）</div>
-                )}
-                {data.alerts.ranked_d_count > 0 && (
-                  <div style={{ fontSize: 14, color: "#fca5a5" }}>ランクD（疑義あり）: {data.alerts.ranked_d_count}件</div>
-                )}
-              </div>
-            </div>
-          )}
-        </>
-      )}
-    </div>
-  );
-}
-
-// ─── タブ: 請求情報 ───────────────────────────────────────────────────────────
-
-function BillingInfoTab({ tenant }: { tenant: TenantDetail }) {
-  return (
-    <div style={{ ...CARD_STYLE, display: "flex", flexDirection: "column", gap: 16 }}>
-      <h3 style={{ fontSize: 15, fontWeight: 700, color: "#d1d5db", margin: 0 }}>💳 請求情報</h3>
-      <div style={{ display: "grid", gap: 10 }}>
-        {[
-          { label: "プラン", value: tenant.plan.toUpperCase() },
-          { label: "課金有効", value: tenant.billing_enabled ? "有効" : "無効" },
-          { label: "無料期間（開始）", value: tenant.billing_free_from ? new Date(tenant.billing_free_from).toLocaleDateString("ja-JP") : "—" },
-          { label: "無料期間（終了）", value: tenant.billing_free_until ? new Date(tenant.billing_free_until).toLocaleDateString("ja-JP") : "—" },
-        ].map(({ label, value }) => (
-          <div key={label} style={{ display: "flex", justifyContent: "space-between", padding: "12px 16px", borderRadius: 8, background: "rgba(255,255,255,0.02)", border: "1px solid #1f2937", fontSize: 14 }}>
-            <span style={{ color: "#9ca3af" }}>{label}</span>
-            <span style={{ color: "#e5e7eb", fontWeight: 600 }}>{value}</span>
-          </div>
-        ))}
-      </div>
-      <p style={{ fontSize: 12, color: "#6b7280", margin: 0 }}>
-        詳細な請求設定はSuper Admin専用の設定タブから変更できます。
-      </p>
-    </div>
-  );
-}
-
-// ─── タブ: 通知設定 ───────────────────────────────────────────────────────────
-
-interface NotificationPref {
-  notification_type: string;
-  email_enabled: boolean;
-  in_app_enabled: boolean;
-  threshold: Record<string, unknown> | null;
-}
-
-const DEFAULT_NOTIFICATION_TYPES = [
-  { type: "ga4_error", label: "GA4接続エラー" },
-  { type: "cv_drop", label: "CV数急減" },
-  { type: "llm_cost_spike", label: "LLMコスト急増" },
-  { type: "weekly_report", label: "週次レポート" },
-];
-
-function NotificationPreferencesTab({ tenantId }: { tenantId: string }) {
-  const [prefs, setPrefs] = useState<Record<string, NotificationPref>>({});
-  const [loading, setLoading] = useState(true);
-  const [saving, setSaving] = useState<string | null>(null);
-  const [toast, setToast] = useState<string | null>(null);
-
-  const showToast = (msg: string) => { setToast(msg); setTimeout(() => setToast(null), 3000); };
-
-  useEffect(() => {
-    const load = async () => {
-      try {
-        const res = await authFetch(`${API_BASE}/v1/admin/tenants/${tenantId}/notification-preferences`);
-        if (!res.ok) return;
-        const json = await res.json() as { preferences: NotificationPref[] };
-        const map: Record<string, NotificationPref> = {};
-        for (const p of json.preferences) map[p.notification_type] = p;
-        setPrefs(map);
-      } finally {
-        setLoading(false);
-      }
-    };
-    void load();
-  }, [tenantId]);
-
-  const handleToggle = async (type: string, field: "email_enabled" | "in_app_enabled") => {
-    const current = prefs[type] ?? { notification_type: type, email_enabled: true, in_app_enabled: true, threshold: null };
-    const updated = { ...current, [field]: !current[field] };
-    setSaving(type);
-    try {
-      const res = await authFetch(`${API_BASE}/v1/admin/tenants/${tenantId}/notification-preferences`, {
-        method: "PUT",
-        body: JSON.stringify({ notification_type: type, email_enabled: updated.email_enabled, in_app_enabled: updated.in_app_enabled }),
-      });
-      if (res.ok) {
-        setPrefs((prev) => ({ ...prev, [type]: updated }));
-        showToast("✅ 保存しました");
-      }
-    } finally {
-      setSaving(null);
-    }
-  };
-
-  if (loading) return <div style={{ color: "#6b7280", textAlign: "center", padding: 32 }}>読み込み中...</div>;
-
-  return (
-    <div style={{ display: "flex", flexDirection: "column", gap: 20 }}>
-      {toast && (
-        <div style={{ padding: "12px 16px", borderRadius: 10, background: "rgba(15,23,42,0.98)", border: "1px solid #22c55e", color: "#4ade80", fontSize: 14, fontWeight: 600 }}>
-          {toast}
-        </div>
-      )}
-      <div style={{ ...CARD_STYLE }}>
-        <h3 style={{ fontSize: 15, fontWeight: 700, color: "#d1d5db", margin: "0 0 16px" }}>🔔 通知設定</h3>
-        <div style={{ display: "grid", gap: 8 }}>
-          {DEFAULT_NOTIFICATION_TYPES.map(({ type, label }) => {
-            const pref = prefs[type] ?? { notification_type: type, email_enabled: true, in_app_enabled: true, threshold: null };
-            const isSavingThis = saving === type;
-            return (
-              <div key={type} style={{ display: "flex", justifyContent: "space-between", alignItems: "center", padding: "12px 16px", borderRadius: 10, background: "rgba(255,255,255,0.02)", border: "1px solid #1f2937" }}>
-                <span style={{ fontSize: 14, color: "#d1d5db", fontWeight: 500 }}>{label}</span>
-                <div style={{ display: "flex", gap: 12 }}>
-                  {(["email_enabled", "in_app_enabled"] as const).map((field) => (
-                    <button
-                      key={field}
-                      type="button"
-                      disabled={isSavingThis}
-                      onClick={() => void handleToggle(type, field)}
-                      style={{
-                        padding: "6px 14px",
-                        minHeight: 32,
-                        borderRadius: 6,
-                        border: pref[field] ? "1px solid #4ade80" : "1px solid #374151",
-                        background: pref[field] ? "rgba(34,197,94,0.15)" : "rgba(0,0,0,0.3)",
-                        color: pref[field] ? "#4ade80" : "#6b7280",
-                        fontSize: 12,
-                        fontWeight: 600,
-                        cursor: isSavingThis ? "not-allowed" : "pointer",
-                        opacity: isSavingThis ? 0.5 : 1,
-                      }}
-                    >
-                      {field === "email_enabled" ? "📧 メール" : "🔔 アプリ内"}
-                    </button>
-                  ))}
-                </div>
-              </div>
-            );
-          })}
-        </div>
-      </div>
-    </div>
   );
 }
